// frontend/src/client/schemas/GlossaryRecordSchema.ts
// This file is autogenerated, do not edit directly.

import {ShortUser} from './ShortUser'

export interface GlossaryRecordSchema {
  id: number
  created_at: string
  updated_at: string
  comment: string | null
  source: string
  target: string
  created_by_user: ShortUser
  glossary_id: number
}